rename accordion item
